Move NonMaxSuppression shape inference into a non_max_suppression submodule
Move the NonMaxSuppression operator, plus its test, from ops.rs into a new
ops/non_max_suppression.rs submodule, matching how this operator is grouped
in src/ops.rs.
